// packages/verify/lib/enums/VerifyLanguages.ts
/**
 * Enum representing supported languages for Verify requests.
 */
export enum VerifyLanguages {
  /**
   * Arabic language (ar-xa).
   */
  ARABIC = 'ar-xa',

  /**
   * Czech language (cs-cz).
   */
  CZECH = 'cs-cz',

  /**
   * Welsh language (cy-cy).
   */
  WELSH = 'cy-cy',

  /**
   * Welsh language, United Kingdom (cy-gb).
   */
  WELSH_UK = 'cy-gb',

  /**
   * Danish language (da-dk).
   */
  DANISH = 'da-dk',

  /**
   * German language (de-de).
   */
  GERMAN = 'de-de',

  /**
   * Greek language (el-gr).
   */
  GREEK = 'el-gr',

  /**
   * English language, Australia (en-au).
   */
  ENGLISH_AUSTRALIA = 'en-au',

  /**
   * English language, United Kingdom (en-gb).
   */
  ENGLISH_UK = 'en-gb',

  /**
   * English language, India (en-in).
   */
  ENGLISH_INDIA = 'en-in',

  /**
   * English language, United States (en-us).
   */
  ENGLISH_US = 'en-us',

  /**
   * Spanish language, Spain (es-es).
   */
  SPANISH_SPAIN = 'es-es',

  /**
   * Spanish language, Mexico (es-mx).
   */
  SPANISH_MEXICO = 'es-mx',

  /**
   * Spanish language, United States (es-us).
   */
  SPANISH_US = 'es-us',

  /**
   * Finnish language (fi-fi).
   */
  FINNISH = 'fi-fi',

  /**
   * Filipino language (fil-ph).
   */
  FILIPINO = 'fil-ph',

  /**
   * French language, Canada (fr-ca).
   */
  FRENCH_CANADA = 'fr-ca',

  /**
   * French language, France (fr-fr).
   */
  FRENCH_FRANCE = 'fr-fr',

  /**
   * Hindi language (hi-in).
   */
  HINDI = 'hi-in',

  /**
   * Hungarian language (hu-hu).
   */
  HUNGARIAN = 'hu-hu',

  /**
   * Indonesian language (id-id).
   */
  INDONESIAN = 'id-id',

  /**
   * Icelandic language (is-is).
   */
  ICELANDIC = 'is-is',

  /**
   * Italian language (it-it).
   */
  ITALIAN = 'it-it',

  /**
   * Japanese language (ja-jp).
   */
  JAPANESE = 'ja-jp',

  /**
   * Korean language (ko-kr).
   */
  KOREAN = 'ko-kr',

  /**
   * Norwegian language (nb-no).
   */
  NORWEGIAN = 'nb-no',

  /**
   * Dutch language (nl-nl).
   */
  DUTCH = 'nl-nl',

  /**
   * Polish language (pl-pl).
   */
  POLISH = 'pl-pl',

  /**
   * Portuguese language, Brazil (pt-br).
   */
  PORTUGUESE_BRAZIL = 'pt-br',

  /**
   * Portuguese language, Portugal (pt-pt).
   */
  PORTUGUESE_PORTUGAL = 'pt-pt',

  /**
   * Romanian language (ro-ro).
   */
  ROMANIAN = 'ro-ro',

  /**
   * Russian language (ru-ru).
   */
  RUSSIAN = 'ru-ru',

  /**
   * Swedish language (sv-se).
   */
  SWEDISH = 'sv-se',

  /**
   * Thai language (th-th).
   */
  THAI = 'th-th',

  /**
   * Turkish language (tr-tr).
   */
  TURKISH = 'tr-tr',

  /**
   * Vietnamese language (vi-vn).
   */
  VIETNAMESE = 'vi-vn',

  /**
   * Yue Chinese Cantonese language (yue-cn).
   */
  YUE_CHINESE_CANTONESE = 'yue-cn',

  /**
   * Chinese language, Mainland (zh-cn).
   */
  CHINESE_MAINLAND = 'zh-cn',

  /**
   * Chinese language, Taiwan (zh-tw).
   */
  CHINESE_TAIWAN = 'zh-tw',
}